refactor: split intelligence into cluster and policy (#734)
#734 proposed merging insights + intelligence + hostclassification into
one `host` module. Reading the packages says that was wrong, and the
right move is a split.

`insights` was already coherent — it is the Adjudicate stage plus the
human annotation feeding it, not a tab. `hostclassification` PRODUCES
classifications while `insights` ADJUDICATES them; those are different
stages, and merging them would have collapsed the seam #512 spent seven
slices establishing.

`intelligence` was the incoherent one: a cluster view and operator
configuration bolted together, with opposite dependency shapes. The
cluster view had zero consumers outside its own controller — a sink. The
configuration is consumed by monitor, and now by subnets and story
through LocalityPolicy — a source. Bundled, anything needing the
configuration dragged a UI view with it.

  com.tracepcap.cluster  — the view behind the Network Cluster tab
  com.tracepcap.policy   — operator-declared rules for interpreting
                           addresses (custom private ranges, IP→org)

Package move only. HTTP paths, route names and class names are all
unchanged, so no contract moves: the regenerated openapi/baseline.json is
byte-identical at 93 paths, which is the check that this stayed internal.

The ArchUnit rule listing feature packages embeds that list in its own
description, so renaming a package changes the rule's identity in the
freeze store. Both the rule and the store key were updated together;
cycles stay at 0 with the two new slices.

Verified against the running stack rather than only by compiling: the
cluster, top-hosts and dns-servers endpoints all return real data for an
analysed capture.
