+++ backend/src/main/java/com/tracepcap/cluster/dto/ClusterGraphResponse.java
@@ -1,0 +1,14 @@
+package com.tracepcap.cluster.dto;
+
+import java.util.List;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class ClusterGraphResponse {
+  private String groupType;
+  private List<ClusterNodeDto> clusters;
+  private List<ClusterEdgeDto> edges;
+  private int hiddenClusters;
+}
