--- backend/src/main/java/com/tracepcap/intelligence/dto/ClusterEdgeDto.java
@@ -1,14 +1,0 @@
-package com.tracepcap.intelligence.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ClusterEdgeDto {
-  private String sourceId;
-  private String targetId;
-  private long totalBytes;
-  private long conversationCount;
-  private String dominantProtocol;
-}
