--- backend/src/main/java/com/tracepcap/intelligence/dto/ClusterGraphResponse.java
@@ -1,14 +1,0 @@
-package com.tracepcap.intelligence.dto;
-
-import java.util.List;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ClusterGraphResponse {
-  private String groupType;
-  private List<ClusterNodeDto> clusters;
-  private List<ClusterEdgeDto> edges;
-  private int hiddenClusters;
-}
