--- backend/src/main/java/com/tracepcap/intelligence/dto/TopHostsResponse.java
@@ -1,11 +1,0 @@
-package com.tracepcap.intelligence.dto;
-
-import java.util.List;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class TopHostsResponse {
-  private List<HostSummaryDto> hosts;
-}
