--- backend/src/main/java/com/tracepcap/intelligence/entity/IpClassification.java
@@ -1,7 +1,0 @@
-package com.tracepcap.intelligence.entity;
-
-/** How a custom range's IPs should be treated: internal (PRIVATE) or external (PUBLIC). */
-public enum IpClassification {
-  PRIVATE,
-  PUBLIC
-}
