+++ backend/src/main/java/com/tracepcap/monitor/service/ChangeDetectionService.java
@@ -12,9 +12,9 @@ import com.tracepcap.analysis.spi.HostClassificationLookup;
 import com.tracepcap.analysis.spi.HostClassificationLookup.HostFacts;
 import com.tracepcap.monitor.spi.LabelStalenessCheck;
 import com.tracepcap.monitor.spi.SnapshotRevalidationHook;
-import com.tracepcap.intelligence.entity.CustomPrivateRangeEntity;
-import com.tracepcap.intelligence.entity.IpClassification;
-import com.tracepcap.intelligence.service.CustomPrivateRangeService;
+import com.tracepcap.policy.entity.CustomPrivateRangeEntity;
+import com.tracepcap.policy.entity.IpClassification;
+import com.tracepcap.policy.service.CustomPrivateRangeService;
 import com.tracepcap.monitor.entity.BaselineDefinitionEntity;
 import com.tracepcap.monitor.entity.NetworkChangeEventEntity;
 import com.tracepcap.monitor.entity.NetworkChangeEventEntity.ChangeType;
