+++ backend/src/main/java/com/tracepcap/policy/dto/IpOrgRuleDto.java
@@ -1,0 +1,16 @@
+package com.tracepcap.policy.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class IpOrgRuleDto {
+  private Long id;
+  private String label;
+  private String cidr;
+}
