+++ backend/src/main/java/com/tracepcap/policy/entity/CustomPrivateRangeEntity.java
@@ -1,0 +1,39 @@
+package com.tracepcap.policy.entity;
+
+import jakarta.persistence.*;
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Entity
+@Table(name = "custom_private_ranges")
+@Getter
+@Setter
+@ToString
+@EqualsAndHashCode(onlyExplicitlyIncluded = true)
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CustomPrivateRangeEntity {
+
+  @EqualsAndHashCode.Include
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(nullable = false, unique = true)
+  private String cidr;
+
+  /** Whether matching IPs are forced to PRIVATE (internal) or PUBLIC (external). */
+  @Enumerated(EnumType.STRING)
+  @Column(nullable = false)
+  private IpClassification classification;
+
+  @Column(name = "created_at", nullable = false)
+  private LocalDateTime createdAt;
+}
