+++ backend/src/main/java/com/tracepcap/policy/entity/IpClassification.java
@@ -1,0 +1,7 @@
+package com.tracepcap.policy.entity;
+
+/** How a custom range's IPs should be treated: internal (PRIVATE) or external (PUBLIC). */
+public enum IpClassification {
+  PRIVATE,
+  PUBLIC
+}
