+++ backend/src/main/java/com/tracepcap/policy/entity/IpOrgRuleEntity.java
@@ -1,0 +1,40 @@
+package com.tracepcap.policy.entity;
+
+import jakarta.persistence.*;
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Entity
+@Table(name = "ip_org_rules")
+@Getter
+@Setter
+@ToString
+@EqualsAndHashCode(onlyExplicitlyIncluded = true)
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class IpOrgRuleEntity {
+
+  @EqualsAndHashCode.Include
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(nullable = false)
+  private String label;
+
+  @Column(nullable = false)
+  private String cidr;
+
+  @Column(name = "prefix_length", nullable = false)
+  private int prefixLength;
+
+  @Column(name = "created_at", nullable = false)
+  private LocalDateTime createdAt;
+}
