+++ backend/src/main/java/com/tracepcap/policy/repository/CustomPrivateRangeRepository.java
@@ -1,0 +1,9 @@
+package com.tracepcap.policy.repository;
+
+import com.tracepcap.policy.entity.CustomPrivateRangeEntity;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomPrivateRangeRepository extends JpaRepository<CustomPrivateRangeEntity, Long> {
+  List<CustomPrivateRangeEntity> findAllByOrderByCreatedAtDesc();
+}
