+++ backend/src/main/java/com/tracepcap/policy/repository/IpOrgRuleRepository.java
@@ -1,0 +1,10 @@
+package com.tracepcap.policy.repository;
+
+import com.tracepcap.policy.entity.IpOrgRuleEntity;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IpOrgRuleRepository extends JpaRepository<IpOrgRuleEntity, Long> {
+  /** Returns all rules ordered by most specific prefix first, then label. */
+  List<IpOrgRuleEntity> findAllByOrderByPrefixLengthDescLabelAsc();
+}
