+++ backend/src/test/java/com/tracepcap/architecture/LayerDependencyTest.java
@@ -139,7 +139,8 @@ class LayerDependencyTest {
                   "com.tracepcap.monitor..",
                   "com.tracepcap.story..",
                   "com.tracepcap.insights..",
-                  "com.tracepcap.intelligence..",
+                  "com.tracepcap.cluster..",
+                  "com.tracepcap.policy..",
                   "com.tracepcap.subnets..",
                   "com.tracepcap.report..",
                   "com.tracepcap.tracer..",
