+++ backend/src/test/resources/archunit_store/stored.rules
@@ -1,6 +1,6 @@
 #
 #Wed Jul 15 23:24:03 SGT 2026
-no\ classes\ that\ reside\ in\ a\ package\ 'com.tracepcap.analysis..'\ should\ depend\ on\ classes\ that\ reside\ in\ any\ package\ ['com.tracepcap.monitor..',\ 'com.tracepcap.story..',\ 'com.tracepcap.insights..',\ 'com.tracepcap.intelligence..',\ 'com.tracepcap.subnets..',\ 'com.tracepcap.report..',\ 'com.tracepcap.tracer..',\ 'com.tracepcap.filter..'],\ because\ \#416\:\ dependency\ direction\ is\ always\ feature\ \u2192\ analysis,\ never\ the\ reverse=10392b59-a713-4415-a8ff-9dde7ddb62d9
+no\ classes\ that\ reside\ in\ a\ package\ 'com.tracepcap.analysis..'\ should\ depend\ on\ classes\ that\ reside\ in\ any\ package\ ['com.tracepcap.monitor..',\ 'com.tracepcap.story..',\ 'com.tracepcap.insights..',\ 'com.tracepcap.cluster..',\ 'com.tracepcap.policy..',\ 'com.tracepcap.subnets..',\ 'com.tracepcap.report..',\ 'com.tracepcap.tracer..',\ 'com.tracepcap.filter..'],\ because\ \#416\:\ dependency\ direction\ is\ always\ feature\ \u2192\ analysis,\ never\ the\ reverse=10392b59-a713-4415-a8ff-9dde7ddb62d9
 no\ classes\ that\ reside\ outside\ of\ package\ 'com.tracepcap.analysis..'\ should\ depend\ on\ classes\ that\ reside\ in\ any\ package\ ['com.tracepcap.analysis.repository..',\ 'com.tracepcap.analysis.entity..'],\ because\ cross-module\ access\ to\ analysis\ must\ go\ through\ analysis.spi\ /\ analysis.dto\ (\#512);\ reaching\ into\ repositories\ or\ entities\ pins\ the\ ingest\ pipeline\ in\ place=68b174ff-3611-45f4-8e40-9e71ef54047f
 no\ classes\ that\ reside\ outside\ of\ packages\ ['com.tracepcap.analysis..',\ 'com.tracepcap.hostlog..',\ 'com.tracepcap.extraction..',\ 'com.tracepcap.file..',\ 'com.tracepcap.conversation..']\ should\ depend\ on\ classes\ that\ have\ fully\ qualified\ name\ 'java.lang.ProcessBuilder',\ because\ only\ Ingest\ plumbing,\ Extract\ (eager\ or\ lazy),\ and\ evidence\ export\ may\ touch\ the\ capture\ (\#512\ stage\ rule\ 4);\ everything\ downstream\ reads\ the\ DB=07124f36-7a7a-48eb-8fee-76ba1915553a
 slices\ matching\ 'com.tracepcap.(*)..'\ should\ be\ free\ of\ cycles=0dba057c-8134-4165-a525-9a78d1387496
